requirements.txt
torch>=2.6.0
torchvision>=0.21.0
matplotlib
seaborn
scikit-learn
numpy
pandas
Pillow
